refactor: migrate backend to Java Spring Boot and remove obsolete documentation and scripts
