--- CHECKLIST.md
@@ -1,272 +1,0 @@
-# Admin Dashboard - Implementation Checklist
-
-## ✅ Backend Implementation (Complete)
-
-### Models Created
-- [x] Coordinator.java - Full coordinator entity
-- [x] User.java - User entity with verification
-- [x] Product.java - Marketplace product entity  
-- [x] Transaction.java - Payment transaction entity
-
-### DTOs Created
-- [x] CoordinatorRequestDTO.java - Registration request
-- [x] CoordinatorResponseDTO.java - Coordinator response
-- [x] UserResponseDTO.java - User response
-- [x] UserStatsDTO.java - User statistics
-- [x] ProductResponseDTO.java - Product response
-- [x] TransactionResponseDTO.java - Transaction response
-- [x] DashboardStatsDTO.java - Dashboard statistics
-- [x] PaymentStatsDTO.java - Payment statistics
-
-### Services Implemented
-- [x] CoordinatorService.java - All CRUD operations
-- [x] UserService.java - User management & verification
-- [x] ProductService.java - Product management
-- [x] TransactionService.java - Transaction & revenue
-- [x] DashboardService.java - Dashboard statistics
-- [x] EventService.java - Updated with DTO support
-
-### Controller & Endpoints
-- [x] AdminController.java - 30+ REST endpoints
-- [x] Dashboard stats endpoint
-- [x] Coordinator management endpoints (6)
-- [x] Event management endpoints (4)
-- [x] User management endpoints (9)
-- [x] Product management endpoints (3)
-- [x] Payment management endpoints (3)
-
-### Security Configuration
-- [x] FirebaseAuthFilter.java - Role-based authentication
-- [x] SecurityConfig.java - Admin route protection
-- [x] Custom claims support (admin, coordinator, user)
-- [x] CORS configuration for React frontend
-
-### Documentation
-- [x] ADMIN_API_DOCUMENTATION.md - Complete API reference
-- [x] ADMIN_DASHBOARD_README.md - Setup & architecture guide
-- [x] IMPLEMENTATION_SUMMARY.md - Overview of implementation
-- [x] REACT_QUICK_START.md - Frontend integration guide
-
-## ✅ Feature Requirements (All Implemented)
-
-### Dashboard Section
-- [x] Total events count
-- [x] Total users count
-- [x] Total products count
-- [x] Products sold count
-- [x] 5 recent events display
-
-### Manage Coordinators Section
-- [x] Registration form with validation
-- [x] First name field
-- [x] Last name field
-- [x] Phone number field
-- [x] Email field
-- [x] Degree programme dropdown (15 programmes)
-- [x] View all coordinators list
-- [x] Active/inactive toggle
-- [x] Delete coordinator
-
-### Event Management Section
-- [x] View all events
-- [x] View individual event details
-- [x] Delete events
-- [x] Filter by category (technology, career, culture, sports)
-- [x] Admin has all coordinator access
-
-### User Management Section
-- [x] View all users
-- [x] View university students separately
-- [x] View external users separately
-- [x] Total university students count
-- [x] Total external users count
-- [x] Total pending verification count
-- [x] Active/inactive user toggle
-- [x] Delete users
-- [x] Filter by user type
-- [x] Verify university students
-- [x] View uploaded ID images for verification
-
-### Product Management Section
-- [x] View all products
-- [x] View product details
-- [x] Delete products
-
-### Payment Section
-- [x] Ticket revenue calculation
-- [x] Marketplace revenue calculation
-- [x] Total revenue display
-- [x] Recent transaction details
-- [x] All transactions list
-
-## 📋 Database Collections
-
-### Firestore Collections Used
-- [x] coordinators - Coordinator profiles
-- [x] users - User accounts
-- [x] events - Event listings
-- [x] products - Marketplace products
-- [x] transactions - Payment records
-
-## 🔒 Security Features
-
-### Authentication & Authorization
-- [x] Firebase JWT token validation
-- [x] Role-based access control (ADMIN, COORDINATOR, USER)
-- [x] Custom claims support
-- [x] Protected admin routes
-- [x] Unauthorized/Forbidden error handling
-
-### CORS Configuration
-- [x] Configured for localhost:3000 (React)
-- [x] Configured for localhost:4200 (Angular)
-- [x] All HTTP methods supported
-- [x] Credentials allowed
-
-## 📊 Degree Programmes (15 Total)
-
-- [x] Animal Production and Food Technology
-- [x] Export Agriculture
-- [x] Aquatic Resources Technology
-- [x] Tea Technology and Value Addition
-- [x] Computer Science and Technology
-- [x] Industrial Information Technology
-- [x] Science & Technology
-- [x] Mineral Resources and Technology
-- [x] Entrepreneurship & Management Studies
-- [x] Hospitality, Tourism & Events Management
-- [x] Human Resource Development
-- [x] English Language & Applied Linguistics
-- [x] Engineering Technology
-- [x] Biosystems Technology Honours
-- [x] Information and Communication Technology Honours
-
-## 🎯 User Types & Statuses
-
-### User Types
-- [x] UNIVERSITY_STUDENT - Requires verification
-- [x] EXTERNAL_USER - No verification needed
-
-### Verification Status
-- [x] PENDING - Awaiting review
-- [x] VERIFIED - Approved
-- [x] REJECTED - Denied
-
-### Product Status
-- [x] AVAILABLE - For sale
-- [x] SOLD - Already sold
-- [x] DELETED - Removed by admin
-
-### Transaction Types & Status
-- [x] TICKET - Event purchases
-- [x] MARKETPLACE - Product purchases
-- [x] Status: PENDING, COMPLETED, FAILED, REFUNDED
-
-## 🧪 Testing Status
-
-### Code Quality
-- [x] No compilation errors
-- [x] Clean code structure
-- [x] Proper package organization
-- [x] Consistent naming conventions
-- [x] Proper exception handling
-
-### API Testing Readiness
-- [x] All endpoints documented
-- [x] Request/response examples provided
-- [x] Error responses documented
-- [x] Authentication instructions included
-
-## 📝 Documentation Quality
-
-### Developer Documentation
-- [x] API reference with all endpoints
-- [x] Request/response examples
-- [x] Error code documentation
-- [x] Authentication setup guide
-- [x] CORS configuration details
-
-### Setup Documentation
-- [x] Prerequisites listed
-- [x] Firebase configuration steps
-- [x] Build instructions
-- [x] Admin user setup guide
-- [x] Testing instructions
-
-### Frontend Integration
-- [x] React examples provided
-- [x] API service layer example
-- [x] Custom hook examples
-- [x] Component examples
-- [x] Protected route example
-
-## 🚀 Deployment Readiness
-
-### Configuration
-- [x] application.properties configured
-- [x] Firebase service account setup
-- [x] CORS origins configurable
-- [x] Port configuration
-
-### Best Practices
-- [x] Dependency injection used
-- [x] Service layer separation
-- [x] DTO pattern implemented
-- [x] Exception handling in place
-- [x] Security filters configured
-
-## 📦 Files Created Summary
-
-**Total Files Created: 23**
-
-### Java Files (19)
-- Models: 4 files
-- DTOs: 8 files
-- Services: 5 files (1 updated)
-- Controllers: 1 file
-- Security: 1 file (updated)
-
-### Documentation Files (4)
-- ADMIN_API_DOCUMENTATION.md
-- ADMIN_DASHBOARD_README.md
-- IMPLEMENTATION_SUMMARY.md
-- REACT_QUICK_START.md
-
-## ✅ Final Status: COMPLETE
-
-All requirements have been successfully implemented. The backend is:
-- ✅ Fully functional
-- ✅ Well documented
-- ✅ Secure with role-based access
-- ✅ Ready for frontend integration
-- ✅ Production-ready architecture
-
-## 🎉 Next Steps
-
-### For Backend Developer
-1. Test all endpoints using Postman/Thunder Client
-2. Set up Firebase Admin SDK to create admin users
-3. Deploy to production server
-
-### For Frontend Developer
-1. Review REACT_QUICK_START.md
-2. Set up Firebase in React app
-3. Create admin user with custom claims
-4. Implement admin dashboard UI
-5. Connect to API endpoints
-
-### Optional Enhancements
-1. Add input validation (Bean Validation)
-2. Implement pagination
-3. Add search functionality
-4. Create audit logs
-5. Add rate limiting
-6. Implement caching
-7. Add email notifications
-8. Create backup/restore functionality
-
----
-
-**Status Date:** January 28, 2026  
-**Implementation:** Complete ✅  
-**Ready for Frontend Integration:** Yes ✅
