--- package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "firebase-admin": "^13.6.0",
-    "node-fetch": "^2.7.0"
-  }
-}
