+++ scripts/get-admin-token.js
