+++ scripts/set-admin.js
