+++ src/main/java/com/example/campusaura/controller/MarketplaceController.java
@@ -2,7 +2,7 @@ package com.example.campusaura.controller;
 
 import com.example.campusaura.security.Roles;
 import com.example.campusaura.service.UserService;
-import com.example.campusaura.model.entity.User;
+import com.example.campusaura.model.User;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
