+++ src/main/java/com/example/campusaura/service/UserService.java
@@ -1,6 +1,6 @@
 package com.example.campusaura.service;
 
-import com.example.campusaura.model.entity.User;
+import com.example.campusaura.model.User;
 
 /**
  * Service interface for User business logic.
